double dottest
Former-commit-id: c42faa3bf40dbab83c298136ce8046936ad36c93

+++ system/seismic/Testveltran.c
@@ -5,7 +5,7 @@
 #include "veltran.h"
 
 int main(void) {
-    float dot1[2], dot2[2];
+    double dot1[2], dot2[2];
     static bool pull=false;
     static int nx=10, ns=20, nt=100; 
     static float x0=0, dx=0.01, s0=0, ds=0.001, t0=0, dt=0.001, s1=0;
@@ -14,8 +14,8 @@ int main(void) {
     sf_dot_test(veltran_lop, nt*ns, nt*nx, dot1, dot2);
     veltran_close();
 
-    printf ("%12.8f ? %12.8f\n",dot1[0],dot1[1]);
-    printf ("%12.8f ? %12.8f\n",dot2[0],dot2[1]);
+    printf ("%12.3f ? %12.3f\n",dot1[0],dot1[1]);
+    printf ("%12.3f ? %12.3f\n",dot2[0],dot2[1]);
 
     exit(0);
 }
